fix: تبدیل خودکار وزن kg به گرم در نمایش - بدون نیاز به migration
وزن‌های قبلی در دیتابیس به صورت kg هستن (مثلا 2.45)
با استفاده از toGrams() خودکار تشخیص میده و ضرب در 1000 میکنه
- cast وزن از integer به float (جلوگیری از truncate)
- accessor های total_weight_grams و actual_weight_grams
- weight_grams برای آیتم‌های سفارش
- تمام ویوها آپدیت شدن

+++ Modules/Warehouse/Models/WarehouseProduct.php
@@ -13,7 +13,7 @@ class WarehouseProduct extends Model
     ];
 
     protected $casts = [
-        'weight' => 'integer',
+        'weight' => 'float',
         'price' => 'decimal:0',
         'wc_product_id' => 'integer',
         'parent_id' => 'integer',
